pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-digital-assistant"
version = "1.0.0"
description = "Professional AI Digital Assistant with Gemini ADK"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "google-adk>=1.19.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "litellm>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.5.0",
    "playwright>=1.44.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
assistant = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
